Create RootView component to simplify App
