+++ src/model/enum.ts
@@ -1,5 +1,5 @@
 // eslint-disable-next-line import/prefer-default-export
 export enum Theme {
-  Light = 'light',
-  Dark = 'dark',
+  Light = 'Light',
+  Dark = 'Dark',
 }
